requirements.txt
flask
speech_recognition
pydub
googletrans
fpdf
python-docx